fix: Fix review comment
