+++ pages/src/main/java/uk/gov/android/ui/pages/instructions/Instructions.kt
@@ -20,31 +20,28 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.Dp
+import uk.gov.android.ui.components.ButtonParameters
+import uk.gov.android.ui.components.ButtonType
+import uk.gov.android.ui.components.GdsButton
+import uk.gov.android.ui.components.GdsHeading
 import uk.gov.android.ui.components.GdsHelpText
 import uk.gov.android.ui.components.GdsVectorImage
+import uk.gov.android.ui.components.HeadingParameters
+import uk.gov.android.ui.components.HeadingSize
 import uk.gov.android.ui.components.HelpTextParameters
 import uk.gov.android.ui.components.VectorImageParameters
 import uk.gov.android.ui.components.content.ContentParameters
 import uk.gov.android.ui.components.content.GdsContent
 import uk.gov.android.ui.components.content.GdsContentText
 import uk.gov.android.ui.components.images.icon.IconParameters
-import uk.gov.android.ui.components.m3.ButtonParametersM3
-import uk.gov.android.ui.components.m3.ButtonTypeM3
-import uk.gov.android.ui.components.m3.GdsButtonM3
-import uk.gov.android.ui.components.m3.HeadingParametersM3
-import uk.gov.android.ui.components.m3.HeadingSizeM3
 import uk.gov.android.ui.pages.R
 import uk.gov.android.ui.pages.brp.BrpInstructionsContentSection
-import uk.gov.android.ui.theme.m3.GdsTheme
+import uk.gov.android.ui.theme.GdsTheme
 import uk.gov.android.ui.theme.mediumPadding
 import uk.gov.android.ui.theme.smallPadding
 import uk.gov.android.ui.theme.xsmallPadding
 
-sealed class Instructions {
-    abstract val generate: @Composable () -> Unit
-}
-
-data class InstructionsParametersM3(
+data class InstructionsParameters(
     @DrawableRes
     var topIcon: Int? = null,
     var topIconScale: ContentScale = ContentScale.Fit,
@@ -61,7 +58,7 @@ data class InstructionsParametersM3(
     val titleAlign: TextAlign = TextAlign.Start,
     val titleBottomPadding: Dp = mediumPadding,
     val helpTextParameters: HelpTextParameters? = null,
-    val buttonParameters: List<ButtonParametersM3>? = null
+    val buttonParameters: List<ButtonParameters>? = null
 ) : Instructions() {
     override val generate: @Composable () -> Unit
         get() = {
@@ -76,8 +73,8 @@ data class InstructionsParametersM3(
                         ),
                     verticalArrangement = Arrangement.SpaceBetween
                 ) {
-                    Content(instructionsParameters = this@InstructionsParametersM3)
-                    Buttons(instructionsParameters = this@InstructionsParametersM3)
+                    Content(instructionsParameters = this@InstructionsParameters)
+                    Buttons(instructionsParameters = this@InstructionsParameters)
                 }
             }
         }
@@ -86,7 +83,7 @@ data class InstructionsParametersM3(
 @Composable
 @Suppress("LongMethod")
 internal fun Content(
-    instructionsParameters: InstructionsParametersM3
+    instructionsParameters: InstructionsParameters
 ) {
     instructionsParameters.apply {
         Column(
@@ -101,18 +98,20 @@ internal fun Content(
                     )
                 )
             }
-            HeadingParametersM3(
-                modifier = Modifier
-                    .fillMaxWidth(),
-                size = HeadingSizeM3.H1(),
-                text = title,
-                textAlign = titleAlign,
-                padding = PaddingValues(
-                    end = smallPadding,
-                    start = smallPadding,
-                    bottom = titleBottomPadding
+            GdsHeading(
+                headingParameters = HeadingParameters(
+                    modifier = Modifier
+                        .fillMaxWidth(),
+                    size = HeadingSize.H1(),
+                    text = title,
+                    textAlign = titleAlign,
+                    padding = PaddingValues(
+                        end = smallPadding,
+                        start = smallPadding,
+                        bottom = titleBottomPadding
+                    )
                 )
-            ).generate()
+            )
             GdsContent(
                 contentParameters = ContentParameters(
                     modifier = Modifier
@@ -152,7 +151,7 @@ internal fun Content(
 
 @Composable
 internal fun Buttons(
-    instructionsParameters: InstructionsParametersM3
+    instructionsParameters: InstructionsParameters
 ) {
     instructionsParameters.buttonParameters?.let {
         Column(
@@ -164,15 +163,15 @@ internal fun Buttons(
                 )
         ) {
             it.forEach { parameters ->
-                GdsButtonM3(parameters)
+                GdsButton(parameters)
             }
         }
     }
 }
 
-class InstructionsM3Provider : PreviewParameterProvider<InstructionsParametersM3> {
-    override val values: Sequence<InstructionsParametersM3> = sequenceOf(
-        InstructionsParametersM3(
+class InstructionsProvider : PreviewParameterProvider<InstructionsParameters> {
+    override val values: Sequence<InstructionsParameters> = sequenceOf(
+        InstructionsParameters(
             topIcon = R.drawable.ic_photo_camera,
             title = R.string.preview__BrpInstructions__title,
             content = listOf(
@@ -198,17 +197,17 @@ class InstructionsM3Provider : PreviewParameterProvider<InstructionsParametersM3
                 )
             ),
             buttonParameters = listOf(
-                ButtonParametersM3(
-                    buttonType = ButtonTypeM3.PRIMARY(),
+                ButtonParameters(
+                    buttonType = ButtonType.PRIMARY(),
                     text = R.string.preview__BrpInstructions__primary_button,
                     modifier = Modifier
                         .fillMaxWidth()
                         .padding(bottom = xsmallPadding),
                     onClick = {}
                 ),
-                ButtonParametersM3(
-                    buttonType = ButtonTypeM3.ICON(
-                        buttonType = ButtonTypeM3.SECONDARY(),
+                ButtonParameters(
+                    buttonType = ButtonType.ICON(
+                        buttonType = ButtonType.SECONDARY(),
                         iconParameters = IconParameters(
                             image = R.drawable.ic_external_site,
                             description = R.string.externalSite
@@ -238,8 +237,8 @@ class InstructionsM3Provider : PreviewParameterProvider<InstructionsParametersM3
 )
 @Composable
 fun Preview(
-    @PreviewParameter(InstructionsM3Provider::class)
-    parameters: InstructionsParametersM3
+    @PreviewParameter(InstructionsProvider::class)
+    parameters: InstructionsParameters
 ) {
     parameters.generate()
 }
